moved to core/optimisation submodule
